Move models to `GraphDetails.kt`
